# docs/openapi.yaml
openapi: 3.1.0
info:
  title: QwenBridge API
  description: |
    OpenAI-compatible reverse proxy for Qwen models with account rotation,
    tool calling, multimodal upload, and streaming support.
  version: 1.0.0
  contact:
    name: QwenBridge
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/models:
    get:
      operationId: listModels
      summary: Lists the currently available models
      description: |
        OpenAI list format by default. When `anthropic-version` is present,
        returns Anthropic-compatible model objects (same underlying catalog,
        including the synthetic `-fast` variant for every model).
      tags:
        - Models
      parameters:
        - name: If-None-Match
          in: header
          schema:
            type: string
          description: ETag value for conditional GET (OpenAI format only). Returns 304 if unchanged.
        - name: anthropic-version
          in: header
          schema:
            type: string
          description: When set, response uses Anthropic model list shape.
      responses:
        "200":
          description: List of models
          headers:
            ETag:
              schema:
                type: string
              description: MD5-based entity tag (OpenAI format)
            Cache-Control:
              schema:
                type: string
              description: Cache policy (public, max-age=3600)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelList"
        "304":
          description: Not modified (ETag match)
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/models/{model}:
    get:
      operationId: retrieveModel
      summary: Retrieves a model instance by ID
      tags:
        - Models
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
          description: The ID of the model to retrieve
        - name: anthropic-version
          in: header
          schema:
            type: string
          description: When set, response uses Anthropic model object shape.
      responses:
        "200":
          description: The model object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "OpenAI-style API key: `Authorization: Bearer <key>`"
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: "Anthropic-style API key header (same secret as `API_KEY`)"
  schemas:
...
    ModelList:
      type: object
      required:
        - object
        - data
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: "#/components/schemas/Model"
    Model:
      type: object
      required:
        - id
        - object
      properties:
        id:
          type: string
          example: qwen3-max
        object:
          type: string
          enum: [model]
        created:
          type: integer
          format: int64
        owned_by:
          type: string
          example: qwen
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
            - type
            - code
          properties:
            message:
              type: string
              description: Human-readable error description
            type:
              type: string
              description: Error category
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found_error
                - rate_limit_error
                - upstream_error
                - upstream_timeout_error
                - service_unavailable_error
                - internal_error
            code:
              type: [string, "null"]
              description: Machine-readable error code
            param:
              type: [string, "null"]
              description: Request parameter associated with the error, if any
  responses:
...
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              message: "Invalid API key"
              type: authentication_error
              code: invalid_api_key
              param: null
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              message: "Not found"
              type: not_found_error
              code: not_found
              param: null
...